pyproject.toml
[project]
name = "symposium-ai"
version = "0.1.0"
description = "Multi-LLM debate arena (Gemini + Claude + GPT)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.7",
  "pydantic-settings>=2.2",
  "python-dotenv>=1.0",
  "openai>=1.40",
  "anthropic>=0.34",
  "google-genai>=0.3.0",
  "rich>=13.7",
  "typer>=0.12",
]

[project.scripts]
arena = "arena.cli:app"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["arena"]
